+++ src/client/AgendaItem.jsx
@@ -11,19 +11,24 @@
 import React from 'react';
 import ListItem from '@mui/material/ListItem';
 import ListItemText from '@mui/material/ListItemText';
+import { format } from 'date-fns';
+import { useTranslation } from './contexts/LocalisationContext.jsx';
 
 export const AgendaItem = ({ agendaItem }) => {
+    // get translation context
+    const { t, dfnsLocale } = useTranslation();
 
     // render jsx
     return (
         <ListItem
             sx={{
                 backgroundColor: agendaItem.calendar.bgColor,
+                mb: 1,
             }}
         >
             <ListItemText
-                primary={`${agendaItem.from} - ${agendaItem.to} ${agendaItem.title}`}
-                secondary={`${agendaItem.text} (${agendaItem.location})`}
+                primary={`${agendaItem.title}${agendaItem.location !== null ? `, ${agendaItem.location}` : ''}${agendaItem.text !== null ? ` (${agendaItem.text})` : ''}`}
+                secondary={`${format(agendaItem.start, 'HH:mm', dfnsLocale)} - ${format(agendaItem.end, 'HH:mm', dfnsLocale)} ${t('o\'clock')}`}
             />
         </ListItem>
     );