+++ src/client/AgendaSection.jsx
@@ -10,13 +10,16 @@
 
 import React from 'react';
 import ListSubheader from '@mui/material/ListSubheader';
-import { format } from 'date-fns';
+import {
+    format,
+    isSameDay,
+} from 'date-fns';
 import { AgendaItem } from './AgendaItem.jsx';
 import { useTranslation } from './contexts/LocalisationContext.jsx';
 
 export const AgendaSection = ({ agendaSection }) => {
     // get translation context
-    const { dfnsLocale } = useTranslation();
+    const { t, dfnsLocale } = useTranslation();
 
     // prepare date-fns options
     const dfnsOptions = {
@@ -27,7 +30,7 @@ export const AgendaSection = ({ agendaSection }) => {
     return (
         <li>
             <ul>
-                <ListSubheader>{format(agendaSection.day, 'EEEE, d. MMMM yyyy', dfnsOptions)}</ListSubheader>
+                <ListSubheader>{isSameDay(agendaSection.day, new Date()) ? t('Today') : format(agendaSection.day, 'EEEE, d. MMMM yyyy', dfnsOptions)}</ListSubheader>
                 {agendaSection.items.map((agendaItem, index) => <AgendaItem key={`item-${index}`} agendaItem={agendaItem} />)}
             </ul>
         </li>