+++ src/server/api/v1/agenda.js
@@ -8,42 +8,25 @@
  */
 'use strict';
 
+import { getAgenda } from '../../Agenda.js';
+
 /**
  * get the appointments from dav and return as agenda
  * @param {Request} req express request object
  * @param {Response} res express response object
  * @returns {Object} the api result
  */
-export const agenda = (req, res) => {
+export const agenda = async (req, res) => {
+
+    // get calendar settings
+    const result = await req.app.locals.redis.json.get('agendadisplay:settings', { path: '$' });
+    const calendars = result[0].calendars;
+    const agendaLength = result[0].agendaLength;
+
+    // get agenda
+    const agenda = await getAgenda(calendars, agendaLength, req.app.locals.redis);
+
     return({
-        agenda: [
-            {
-                day: new Date('2024-05-19'),
-                items: [
-                    {
-                        calendar: {
-                            name: 'Calendar 1',
-                            bgColor: 'red',
-                        },
-                        from: '9:00',
-                        to: '10:00',
-                        title: 'Test Item 1',
-                        text: 'Test Test Test Test Test Test Test Test',
-                        location: 'Location',
-                    },
-                    {
-                        calendar: {
-                            name: 'Calendar 2',
-                            bgColor: 'blue',
-                        },
-                        from: '13:00',
-                        to: '20:00',
-                        title: 'Test Item 2',
-                        text: 'Test Test Test Test Test Test Test Test',
-                        location: '',
-                    },
-                ],
-            },
-        ],
+        agenda: agenda,
     });
 }